Test: Refactor tests
