+++ tests/Unit/JobInitializers/AlwaysFreshInitializerTest.php
@@ -9,6 +9,9 @@ use Illuminate\Queue\Events\JobProcessing;
 use Mockery;
 use Mpyw\LaravelCachedDatabaseStickiness\JobInitializers\AlwaysFreshInitializer;
 use Mpyw\LaravelCachedDatabaseStickiness\StickinessManager;
+use Mpyw\LaravelCachedDatabaseStickiness\Tests\FreshJob;
+use Mpyw\LaravelCachedDatabaseStickiness\Tests\GeneralJob;
+use Mpyw\LaravelCachedDatabaseStickiness\Tests\ModifiedJob;
 use Orchestra\Testbench\TestCase;
 
 class AlwaysFreshInitializerTest extends TestCase
@@ -51,12 +54,13 @@ class AlwaysFreshInitializerTest extends TestCase
 
     public function testGeneralJob(): void
     {
-        $job = new class() {
-        };
+        $job = new GeneralJob();
 
         $this->job->shouldReceive('payload')->once()->andReturn([
+            'job' => 'Illuminate\Queue\CallQueuedHandler@call',
             'data' => [
                 'commandName' => get_class($job),
+                'command' => serialize($job),
             ],
         ]);
         $this->db->shouldReceive('getConnections')->once()->andReturn([$this->connection]);
@@ -67,12 +71,13 @@ class AlwaysFreshInitializerTest extends TestCase
 
     public function testFreshJob(): void
     {
-        $job = new class() implements \Mpyw\LaravelCachedDatabaseStickiness\Jobs\ShouldAssumeFresh {
-        };
+        $job = new FreshJob();
 
         $this->job->shouldReceive('payload')->once()->andReturn([
+            'job' => 'Illuminate\Queue\CallQueuedHandler@call',
             'data' => [
                 'commandName' => get_class($job),
+                'command' => serialize($job),
             ],
         ]);
         $this->db->shouldReceive('getConnections')->once()->andReturn([$this->connection]);
@@ -83,12 +88,13 @@ class AlwaysFreshInitializerTest extends TestCase
 
     public function testModifiedJob(): void
     {
-        $job = new class() implements \Mpyw\LaravelCachedDatabaseStickiness\Jobs\ShouldAssumeModified {
-        };
+        $job = new ModifiedJob();
 
         $this->job->shouldReceive('payload')->once()->andReturn([
+            'job' => 'Illuminate\Queue\CallQueuedHandler@call',
             'data' => [
                 'commandName' => get_class($job),
+                'command' => serialize($job),
             ],
         ]);
         $this->db->shouldReceive('getConnections')->once()->andReturn([$this->connection]);
@@ -100,8 +106,10 @@ class AlwaysFreshInitializerTest extends TestCase
     public function testBrokenCommandNameJob(): void
     {
         $this->job->shouldReceive('payload')->once()->andReturn([
+            'job' => 'Illuminate\Queue\CallQueuedHandler@call',
             'data' => [
                 'commandName' => ['foo'],
+                'command' => [],
             ],
         ]);
         $this->db->shouldReceive('getConnections')->once()->andReturn([$this->connection]);
@@ -113,6 +121,7 @@ class AlwaysFreshInitializerTest extends TestCase
     public function testMissingCommandNameJob(): void
     {
         $this->job->shouldReceive('payload')->once()->andReturn([
+            'job' => 'Illuminate\Queue\CallQueuedHandler@call',
             'data' => [],
         ]);
         $this->db->shouldReceive('getConnections')->once()->andReturn([$this->connection]);
